Added Fade Functionality
Added an appropriate response to the bri_inc command.  It will now respond the same way that a Hue Bulb does.  So, for example, with a hue dimmer switch, if your light is at 100% and you press the button to fade down the light, the light will fade down until the button is released and then stop.